move parse_spec to util.spec
