+++ mitmproxy/addons/modifyheaders.py
@@ -5,32 +5,7 @@ from pathlib import Path
 from mitmproxy import ctx, exceptions, flowfilter, http
 from mitmproxy.net.http import Headers
 from mitmproxy.utils import strutils
-
-
-def _match_all(flow) -> bool:
-    return True
-
-
-def parse_spec(option: str) -> typing.Tuple[flowfilter.TFilter, str, str]:
-    """
-    Parse strings in the following format:
-
-        [/flow-filter]/subject/replacement
-
-    """
-    sep, rem = option[0], option[1:]
-    parts = rem.split(sep, 2)
-    if len(parts) == 2:
-        subject, replacement = parts
-        return _match_all, subject, replacement
-    elif len(parts) == 3:
-        patt, subject, replacement = parts
-        flow_filter = flowfilter.parse(patt)
-        if not flow_filter:
-            raise ValueError(f"Invalid filter pattern: {patt}")
-        return flow_filter, subject, replacement
-    else:
-        raise ValueError("Invalid number of parameters (2 or 3 are expected)")
+from mitmproxy.utils.spec import parse_spec
 
 
 class ModifySpec(typing.NamedTuple):
